# docs/swagger.yaml
basePath: /api
...
host: localhost:8080
info:
  contact: {}
  description: API untuk manajemen tugas dan proyek
  title: Gintugas API
  version: "1.0"
paths:
  /api/admin/users:
    get:
      description: Mendapatkan daftar semua users (hanya admin)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get semua users
      tags:
      - admin
  /api/admin/users/{id}:
    delete:
      description: Hapus user (hanya admin)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - admin
    get:
      description: Mendapatkan detail user berdasarkan ID (hanya admin)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: Update data user (hanya admin)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Data user
        in: body
        name: input
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - admin
...
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Melakukan login dan mendapatkan token JWT
      parameters:
      - description: Kredensial login { \
        in: body
        name: input
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Login success
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Login user
      tags:
      - auth
  /api/auth/logout:
    post:
      description: Melakukan logout dan invalidate token
      produces:
      - application/json
      responses:
        "200":
          description: Logout success
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Mendaftarkan user baru ke sistem
      parameters:
      - description: Data registrasi { \
        in: body
        name: input
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Register success
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Register user baru
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Authorization header menggunakan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"